requirements.txt
streamlit
openai
pytest
python-dotenv
